requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.0
supabase==1.0.3
pydantic==2.4.2
python-dotenv==1.0.0
uuid==1.30
aiofiles==23.2.1
pydub==0.25.1
ffmpeg-python==0.2.0
